The app seems to be working!
* You can start and stop scanning
* The second page shows the regions (that are currently hard-wired into the app) that we’re searching for
* If you enter any of the regions, a pair of ruby slippers shows up next to that region name
* If you tap any region name, it shows you the details of that region (whether you’re in or out of it, the uuid, last time it saw a beacon for that region, the major/minor for the beacon, etc.)

And it seems to detect the RPi beacon well.  Yay!  It seems to take about 30 seconds for iOS to decide that you are outside the region (e.g., if I shut off the beacon, iOS tells me that we’re “outside” the region about 30 seconds later).  This is fine for our purposes.